+++ turbo/apps/host-worker/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## [1.5.34](https://github.com/okou-ai/okou/compare/host-worker-v1.5.33...host-worker-v1.5.34) (2026-09-23)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @okouai/api-contracts bumped to 1.497.1
+
 ## [1.5.33](https://github.com/okou-ai/okou/compare/host-worker-v1.5.32...host-worker-v1.5.33) (2026-09-23)